requirements.txt
# Base dependencies
itchat-uos==1.5.0.dev0
wxpy==0.3.9.8

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pypng==0.20220715.0
pillow==10.2.0

# Logging
loguru==0.7.2

# Environment
python-dotenv==1.0.1

# Development dependencies (optional)
# pytest==8.0.0
# black==24.1.1